+++ sentry/src/main/java/io/sentry/TypeCheckHint.java
@@ -141,6 +141,6 @@ public final class TypeCheckHint {
   /** Used for Ktor Request breadcrumbs. */
   public static final String KTOR_CLIENT_REQUEST = "ktorClient:request";
 
-  /** Used for Session Replay frame bitmaps in the ReplaySnapshotObserver callback. */
+  /** Used for Session Replay frame bitmaps in the ReplayFrameObserver callback. */
   public static final String REPLAY_FRAME_BITMAP = "replay:frameBitmap";
 }